requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
backports-functools-lru-cache=1.3=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py39h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=41.0.3=py39hdda0065_0
cudatoolkit=11.3.1=h2bc3f7f_2
datasets=2.15.0=pypi_0
dawg-python=0.7.2=pypi_0
dill=0.3.6=pypi_0
docopt=0.6.2=pypi_0
et-xmlfile=1.1.0=pypi_0
evaluate=0.4.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=pypi_0
fsspec=2023.10.0=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
greenlet=3.0.1=pypi_0
huggingface-hub=0.19.4=pypi_0
idna=3.4=py39h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
intervaltree=2.1.0=pypi_0
jdcal=1.4.1=pypi_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numpy=1.26.0=py39h5f9d8c6_0
numpy-base=1.26.0=py39hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=2.4.11=pypi_0
openssl=3.0.12=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.1.3=pypi_0
pillow=10.0.1=py39ha6cbd5a_0
pip=23.3=py39h06a4308_0
protobuf=3.20.1=pypi_0
pyarrow=14.0.1=pypi_0
pyarrow-hotfix=0.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pymorphy2=0.8=pypi_0
pymorphy2-dicts=2.4.393442.3710985=pypi_0
pyopenssl=23.2.0=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch=1.12.1=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
records=0.5.3=pypi_0
regex=2023.10.3=pypi_0
requests=2.31.0=py39h06a4308_0
responses=0.18.0=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sql-metadata=2.10.0=pypi_0
sqlalchemy=2.0.23=pypi_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=pypi_0
tablib=3.5.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
torchaudio=0.12.1=py39_cu113
torchvision=0.13.1=py39_cu113
tqdm=4.66.1=pypi_0
transformers=4.20.1=pypi_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023.3=pypi_0
urllib3=1.26.18=py39h06a4308_0
wheel=0.41.2=py39h06a4308_0
xxhash=3.4.1=pypi_0
xz=5.4.2=h5eee18b_0
yargy=0.11.0=pypi_0
yarl=1.9.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
